Refactoring and moving API and create Building Script to shared
