+++ Assets/Shared/Scripts/Building.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 6c6679a2813a3489d9c284e3feeb5cca
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
