+++ Assets/Shared/Scripts/Building/CreateBuilding_report.md.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: a89032fedfba84d38897bf049eee9180
+TextScriptImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
